refactor: farmer videos
