+++ libs/data/farmer_content/data/content/4_what_are_the_options.ts
@@ -1,15 +1,15 @@
 import { marker as translateMarker } from '@biesbjerg/ngx-translate-extract-marker';
 
-import { PICSA_FARMER_VIDEO_RESOURCES } from '@picsa/resources/src/app/data/picsa/farmer-videos';
 import { IFarmerContent, IFarmerContentStep } from '../../types';
 import { TOOLS_DATA_HASHMAP } from '../tools';
+import { PICSA_FARMER_VIDEOS_HASHMAP } from '@picsa/data/resources';
 
 const { options } = TOOLS_DATA_HASHMAP;
 
 const steps: IFarmerContentStep[] = [
   {
     type: 'video',
-    resource: PICSA_FARMER_VIDEO_RESOURCES.mw_ny['360p'].options,
+    video: PICSA_FARMER_VIDEOS_HASHMAP.options,
   },
 ];
